+++ src/index.ts
@@ -1,2 +1,3 @@
+export * from './parameters';
 export * from './shared';
 export * from './user-accounts';