+++ src/parameters/request-response/Converter.ts
@@ -1,0 +1,10 @@
+export class Converter {
+
+    public static toGroup(parameterGroups: string[]): string {
+        if (!parameterGroups || parameterGroups.length === 0) {
+            return '';
+        }
+
+        return `&group=${parameterGroups.join(',')}`;
+    }
+}