+++ src/user-accounts/index.ts
@@ -1,4 +1,4 @@
 export * from './AccessRights';
 export * from './User';
-export * from './UserAlreadyExistsError';
 export * from './UserAccounts';
+export * from './UserAlreadyExistsError';