+++ src/user-accounts/request-response/UpdateUserRequest.ts
@@ -1,5 +1,6 @@
 import { Connection, User } from './../..';
-import { Request } from './Request';
+import { Request } from './../../shared/Request';
+import { Converter } from './Converter';
 import { UpdateUserResponse } from './UpdateUserResponse';
 
 export class UpdateUserRequest extends Request {
@@ -16,6 +17,6 @@ export class UpdateUserRequest extends Request {
     }
 
     public get url(): string {
-        return `${this.connection.url}/axis-cgi/pwdgrp.cgi?action=update&user=${this.user.name}&pwd=${this.user.password}&grp=users&sgrp=${this.toUserGroups(this.user.accessRights, this.user.ptz)}`;
+        return `${this.connection.url}/axis-cgi/pwdgrp.cgi?action=update&user=${this.user.name}&pwd=${this.user.password}&grp=users&sgrp=${Converter.toUserGroups(this.user.accessRights, this.user.ptz)}`;
     }
 }