+++ src/user-accounts/request-response/UpdateUserResponse.ts
@@ -1,5 +1,5 @@
 import { UnknownError } from './../..';
-import { Response } from './Response';
+import { Response } from './../../shared/Response';
 
 export class UpdateUserResponse extends Response {
     private static readonly SuccessResponse = /Modified account .*\./;