+++ test/user-accounts/User.spec.ts
@@ -1,62 +1,64 @@
-import { expect } from 'chai';
+import * as chai from 'chai';
 
 import { AccessRights, User } from './../../src';
 import { ExpectationError } from './../../src/shared/expectations/ExpectationError';
 import { Generate } from './Generate';
 
-describe('Users', () => {
+chai.should();
 
-    describe('#ctor(name, ...)', () => {
+describe('users', function() {
 
-        it('should throw exception if to short', () => {
+    describe('#ctor(name, ...)', function() {
+
+        it('should throw exception if to short', function() {
             // Act
             const fn = () => new User(Generate.string(0), 'secret', AccessRights.Viewer, false);
 
             // Assert
-            expect(fn).to.throw(ExpectationError);
+            fn.should.throw(ExpectationError);
         });
 
-        it('should throw exception if to long', () => {
+        it('should throw exception if to long', function() {
             // Act
             const fn = () => new User(Generate.string(15), 'secret', AccessRights.Viewer, false);
 
             // Assert
-            expect(fn).to.throw(ExpectationError);
+            fn.should.throw(ExpectationError);
         });
 
-        it('should throw exception if containing unsupported characters', () => {
+        it('should throw exception if containing unsupported characters', function() {
             // Act
             const fn = () => new User('Joe-', 'secret', AccessRights.Viewer, false);
 
             // Assert
-            expect(fn).to.throw(ExpectationError);
+            fn.should.throw(ExpectationError);
         });
     });
 
-    describe('#ctor(..., password, ...)', () => {
+    describe('#ctor(..., password, ...)', function() {
 
-        it('should throw exception if to short', () => {
+        it('should throw exception if to short', function() {
             // Act
             const fn = () => new User('Joe', Generate.string(0), AccessRights.Viewer, false);
 
             // Assert
-            expect(fn).to.throw(ExpectationError);
+            fn.should.throw(ExpectationError);
         });
 
-        it('should throw exception if to long', () => {
+        it('should throw exception if to long', function() {
             // Act
             const fn = () => new User('Joe', Generate.string(65), AccessRights.Viewer, false);
 
             // Assert
-            expect(fn).to.throw(ExpectationError);
+            fn.should.throw(ExpectationError);
         });
 
-        it('should throw exception if containing unsupported characters', () => {
+        it('should throw exception if containing unsupported characters', function() {
             // Act
             const fn = () => new User('Joe', `secret-${String.fromCharCode(31)}`, AccessRights.Viewer, false);
 
             // Assert
-            expect(fn).to.throw(ExpectationError);
+            fn.should.throw(ExpectationError);
         });
     });
 });