+++ test/user-accounts/request-response/AddUserRequest.spec.ts
@@ -1,15 +1,17 @@
-import { expect } from 'chai';
+import * as chai from 'chai';
 
 import { AccessRights, Connection, Protocol, User } from './../../../src';
 import { AddUserRequest } from './../../../src/user-accounts/request-response/AddUserRequest';
 
-describe('Add user request', () => {
+chai.should();
+
+describe('add user request', function() {
 
     const connection = new Connection(Protocol.Http, '1.2.3.4', 80, 'root', 'pass');
 
-    describe('#url', () => {
+    describe('#url', function() {
 
-        it('should return URL when adding user with viewer access', () => {
+        it('should return URL when adding user with viewer access', function() {
             // Arrange
             const user = new User('John', 'password', AccessRights.Viewer, false);
 
@@ -17,10 +19,10 @@ describe('Add user request', () => {
             const request = new AddUserRequest(connection, user);
 
             // Assert
-            expect(request.url).to.equal(`${connection.url}/axis-cgi/pwdgrp.cgi?action=add&user=${user.name}&pwd=${user.password}&grp=users&sgrp=viewer&comment=${user.name}`);
+            request.url.should.equal(`${connection.url}/axis-cgi/pwdgrp.cgi?action=add&user=${user.name}&pwd=${user.password}&grp=users&sgrp=viewer&comment=${user.name}`);
         });
 
-        it('should return URL when adding user with viewer access and PTZ control', () => {
+        it('should return URL when adding user with viewer access and PTZ control', function() {
             // Arrange
             const user = new User('John', 'password', AccessRights.Viewer, true);
 
@@ -28,10 +30,10 @@ describe('Add user request', () => {
             const request = new AddUserRequest(connection, user);
 
             // Assert
-            expect(request.url).to.equal(`${connection.url}/axis-cgi/pwdgrp.cgi?action=add&user=${user.name}&pwd=${user.password}&grp=users&sgrp=viewer:ptz&comment=${user.name}`);
+            request.url.should.equal(`${connection.url}/axis-cgi/pwdgrp.cgi?action=add&user=${user.name}&pwd=${user.password}&grp=users&sgrp=viewer:ptz&comment=${user.name}`);
         });
 
-        it('should return URL when adding user with operator access', () => {
+        it('should return URL when adding user with operator access', function() {
             // Arrange
             const user = new User('John', 'password', AccessRights.Operator, false);
 
@@ -39,10 +41,10 @@ describe('Add user request', () => {
             const request = new AddUserRequest(connection, user);
 
             // Assert
-            expect(request.url).to.equal(`${connection.url}/axis-cgi/pwdgrp.cgi?action=add&user=${user.name}&pwd=${user.password}&grp=users&sgrp=operator:viewer&comment=${user.name}`);
+            request.url.should.equal(`${connection.url}/axis-cgi/pwdgrp.cgi?action=add&user=${user.name}&pwd=${user.password}&grp=users&sgrp=operator:viewer&comment=${user.name}`);
         });
 
-        it('should return URL when adding user with operator access and PTZ control', () => {
+        it('should return URL when adding user with operator access and PTZ control', function() {
             // Arrange
             const user = new User('John', 'password', AccessRights.Operator, true);
 
@@ -50,10 +52,10 @@ describe('Add user request', () => {
             const request = new AddUserRequest(connection, user);
 
             // Assert
-            expect(request.url).to.equal(`${connection.url}/axis-cgi/pwdgrp.cgi?action=add&user=${user.name}&pwd=${user.password}&grp=users&sgrp=operator:viewer:ptz&comment=${user.name}`);
+            request.url.should.equal(`${connection.url}/axis-cgi/pwdgrp.cgi?action=add&user=${user.name}&pwd=${user.password}&grp=users&sgrp=operator:viewer:ptz&comment=${user.name}`);
         });
 
-        it('should return URL when adding user with administrator access', () => {
+        it('should return URL when adding user with administrator access', function() {
             // Arrange
             const user = new User('John', 'password', AccessRights.Administrator, false);
 
@@ -61,10 +63,10 @@ describe('Add user request', () => {
             const request = new AddUserRequest(connection, user);
 
             // Assert
-            expect(request.url).to.equal(`${connection.url}/axis-cgi/pwdgrp.cgi?action=add&user=${user.name}&pwd=${user.password}&grp=users&sgrp=admin:operator:viewer&comment=${user.name}`);
+            request.url.should.equal(`${connection.url}/axis-cgi/pwdgrp.cgi?action=add&user=${user.name}&pwd=${user.password}&grp=users&sgrp=admin:operator:viewer&comment=${user.name}`);
         });
 
-        it('should return URL when adding user with administrator access and PTZ control', () => {
+        it('should return URL when adding user with administrator access and PTZ control', function() {
             // Arrange
             const user = new User('John', 'password', AccessRights.Administrator, true);
 
@@ -72,7 +74,7 @@ describe('Add user request', () => {
             const request = new AddUserRequest(connection, user);
 
             // Assert
-            expect(request.url).to.equal(`${connection.url}/axis-cgi/pwdgrp.cgi?action=add&user=${user.name}&pwd=${user.password}&grp=users&sgrp=admin:operator:viewer:ptz&comment=${user.name}`);
+            request.url.should.equal(`${connection.url}/axis-cgi/pwdgrp.cgi?action=add&user=${user.name}&pwd=${user.password}&grp=users&sgrp=admin:operator:viewer:ptz&comment=${user.name}`);
         });
     });
 });